+++ test/svelte3/integration/locations/locations.multiscripts.svelte
@@ -1,0 +1,10 @@
+<h1>Template header</h1>
+<p>Template text</p>
+
+<script scope="module">
+    let staticVariable = 1;
+</script>
+
+<script>
+    let variable = 1;
+</script>